requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
google-api-python-client==2.160.0
google-auth==2.37.0
python-docx==1.1.2
PyPDF2==3.0.1
python-dotenv==1.0.1
httpx==0.28.1
google-cloud-secret-manager==2.21.1
